+++ tests/test_hypothesis.py
@@ -2,7 +2,9 @@
 Property-based tests using https://warehouse.python.org/project/hypothesis/
 """
 
-from bidict import bidict, orderedbidict
+from bidict import (bidict, loosebidict, looseorderedbidict, orderedbidict,
+                    OrderedBidirectionalMapping)
+from bidict.compat import iteritems, viewitems
 from hypothesis import assume, given, settings
 from hypothesis.strategies import (
     binary, booleans, choices, dictionaries, floats, frozensets, integers,
@@ -17,7 +19,7 @@ settings.load_profile(getenv('HYPOTHESIS_PROFILE', 'default'))
 
 
 def inv(d):
-    return {v: k for (k, v) in d.items()}
+    return {v: k for (k, v) in iteritems(d)}
 
 
 def prune_dup_vals(d):
@@ -35,13 +37,12 @@ def eq_nan(a, b):
     return a == b or both_nan(a, b)
 
 
+mutable_bidict_types = (bidict, loosebidict, looseorderedbidict, orderedbidict)
 mutating_methods_by_arity = {
-    0: (bidict.clear, bidict.popitem, orderedbidict.popitem,),
-    1: (bidict.__delitem__, bidict.pop, bidict.setdefault,
-        orderedbidict.move_to_end,),
-    2: (bidict.__setitem__, bidict.pop, bidict.put, bidict.forceput,
-        bidict.setdefault,),
-    -1: (bidict.update, bidict.forceupdate,),
+    0: ('clear', 'popitem',),
+    1: ('__delitem__', 'pop', 'setdefault', 'move_to_end',),
+    2: ('__setitem__', 'pop', 'put', 'forceput', 'setdefault',),
+    -1: ('update', 'forceupdate',),
 }
 # otherwise data gen. in hypothesis>=1.19 is so slow the health checks fail:
 kw = dict(average_size=2)
@@ -60,7 +61,7 @@ def test_len(d):
 @given(d)
 def test_bidirectional_mappings(d):
     b = bidict(d)
-    for k, v in b.items():
+    for k, v in iteritems(b):
         assert eq_nan(k, b.inv[v])
 
 
@@ -83,23 +84,42 @@ def test_equality(d):
     assert not b.inv != i
 
 
-@given(d, immutable, lists(tuples(immutable, immutable)))
-def test_consistency_after_mutation(d, arg, itemlist):
-    for arity, mms in mutating_methods_by_arity.items():
-        for mm in mms:
-            b = orderedbidict(d) if 'orderedbidict' in repr(mm) else bidict(d)
-            args = []
-            if arity > 0:
-                args.append(arg)
-            if arity > 1:
-                args.append(arg)
-            if arity == -1:  # update and forceupdate
-                args.append(itemlist)
-            assert b == inv(b.inv)
-            assert b.inv == inv(b)
-            try:
-                mm(b, *args)
-            except:
-                pass
-            assert b == inv(b.inv)
-            assert b.inv == inv(b)
+@given(d, immutable, immutable, lists(tuples(immutable, immutable)))
+def test_consistency_after_mutation(d, arg1, arg2, itemlist):
+    for arity, mms in iteritems(mutating_methods_by_arity):
+        for B in mutable_bidict_types:
+            ordered = issubclass(B, OrderedBidirectionalMapping)
+            for mname in mms:
+                mm = getattr(B, mname, None)
+                if not mm:
+                    continue
+                b = B(d)
+                args = []
+                if arity > 0:
+                    args.append(arg1)
+                if arity > 1:
+                    args.append(arg2)
+                if arity == -1:  # update and forceupdate
+                    args.append(itemlist)
+                assert dict(b) == inv(b.inv)
+                assert dict(b.inv) == inv(b)
+                if ordered:
+                    items1 = list(viewitems(b))
+                try:
+                    mm(b, *args)
+                except:
+                    pass
+                assert dict(b) == inv(b.inv)
+                assert dict(b.inv) == inv(b)
+                if ordered and mname != 'move_to_end':
+                    items2 = list(viewitems(b))
+                    common = set(items1) & set(items2)
+                    for i in common:
+                        idx1 = items1.index(i)
+                        idx2 = items2.index(i)
+                        beforei1 = [j for j in items1[:idx1] if j in common]
+                        beforei2 = [j for j in items2[:idx2] if j in common]
+                        assert beforei1 == beforei2
+                        afteri1 = [j for j in items1[idx1+1:] if j in common]
+                        afteri2 = [j for j in items2[idx2+1:] if j in common]
+                        assert afteri1 == afteri2